+++ swing/common-model-tools/src/main/java/is/codion/swing/common/model/tools/loadtest/DefaultLoadTestModel.java
@@ -284,7 +284,7 @@ final class DefaultLoadTestModel<T> implements LoadTestModel<T> {
 	}
 
 	private void bindEvents() {
-		loadTest.addResultListener(counter::addScenarioResults);
+		loadTest.resultObserver().addDataListener(counter::addScenarioResults);
 		loadTest.addShutdownListener(() -> {
 			applicationsRefreshScheduler.stop();
 			chartUpdateScheduler.stop();